requirements.txt
numpy==2.3.4
opencv_python==4.9.0.80
Pillow==12.0.0
PyAutoGUI==0.9.54
pynput==1.7.7
pywin32==305.1
